# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GYM API Documentation",
        "description": "GYM API Documentation for make connections to your application",
        "contact": {
            "name": "Brenbala",
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/genders": {
            "get": {
                "tags": [
                    "Genders"
                ],
                "summary": "Get all genders",
                "description": "Returns a list of all genders.",
                "operationId": "getGenders",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/goals": {
            "get": {
                "tags": [
                    "Goals"
                ],
                "summary": "Get all fitness goals",
                "description": "Returns a list of all fitness goals.",
                "operationId": "getGoals",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get authenticated user details",
                "description": "Returns the details of the authenticated user.",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "put": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update user details",
                "description": "Updates the details of the authenticated user.",
                "operationId": "updateUser",
                "requestBody": {
                    "description": "User details to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "phone",
                                    "profilePicture",
                                    "gender",
                                    "age",
                                    "height",
                                    "weight",
                                    "weightType",
                                    "physicalActivityLevel",
                                    "goals",
                                    "newPassword",
                                    "confirmPassword"
                                ],
                                "properties": {
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "profilePicture": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "integer"
                                    },
                                    "age": {
                                        "type": "integer"
                                    },
                                    "height": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "weight": {
                                        "type": "integer"
                                    },
                                    "weightType": {
                                        "type": "integer"
                                    },
                                    "physicalActivityLevel": {
                                        "type": "integer"
                                    },
                                    "goals": {
                                        "type": "array",
                                        "items": {}
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "minLength": 8
                                    },
                                    "confirmPassword": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User details updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "firstName": {
                                                    "type": "string"
                                                },
                                                "lastName": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "profilePicture": {
                                                    "type": "string"
                                                },
                                                "gender": {
                                                    "type": "integer"
                                                },
                                                "age": {
                                                    "type": "integer"
                                                },
                                                "height": {
                                                    "type": "number",
                                                    "format": "double"
                                                },
                                                "weight": {
                                                    "type": "integer"
                                                },
                                                "weightType": {
                                                    "type": "integer"
                                                },
                                                "physicalActivityLevel": {
                                                    "type": "integer"
                                                },
                                                "goals": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}